modify students

+++ src/main/java/app/server/bean/RatingDetail.java
@@ -7,16 +7,16 @@ package app.server.bean;
  * @Modified By:
  */
 public class RatingDetail {
-    private int score1 = 0;
-    private int score2 = 0;
-    private int score3 = 0;
-    private int score4 = 0;
-    private int score5 = 0;
+    private double score1 = 0;
+    private double score2 = 0;
+    private double score3 = 0;
+    private double score4 = 0;
+    private double score5 = 0;
 
     public RatingDetail() {
     }
 
-    public RatingDetail(int score1, int score2, int score3, int score4, int score5) {
+    public RatingDetail(double score1, double score2, double score3, double score4, double score5) {
         this.score1 = score1;
         this.score2 = score2;
         this.score3 = score3;
@@ -24,43 +24,43 @@ public class RatingDetail {
         this.score5 = score5;
     }
 
-    public int getScore1() {
+    public double getScore1() {
         return score1;
     }
 
-    public void setScore1(int score1) {
+    public void setScore1(double score1) {
         this.score1 = score1;
     }
 
-    public int getScore2() {
+    public double getScore2() {
         return score2;
     }
 
-    public void setScore2(int score2) {
+    public void setScore2(double score2) {
         this.score2 = score2;
     }
 
-    public int getScore3() {
+    public double getScore3() {
         return score3;
     }
 
-    public void setScore3(int score3) {
+    public void setScore3(double score3) {
         this.score3 = score3;
     }
 
-    public int getScore4() {
+    public double getScore4() {
         return score4;
     }
 
-    public void setScore4(int score4) {
+    public void setScore4(double score4) {
         this.score4 = score4;
     }
 
-    public int getScore5() {
+    public double getScore5() {
         return score5;
     }
 
-    public void setScore5(int score5) {
+    public void setScore5(double score5) {
         this.score5 = score5;
     }
